{"types":"C","runs":7}

+++ Problem15.java
@@ -5,7 +5,7 @@ public class Problem15 extends ConsoleProgram {
 	public void run(){
 		int num1 = readInt("Write first whole number ");
 		int num2 = readInt("Write second whole number ");
-		int avg = (num1+num2)/2;
+		double avg = (num1+num2)/2;
 		println("Average of this two numbers is "+ avg);
 	}
